Refactor adaptive brightness handling and improve debug configuration
- Moved adaptive brightness logic to a new core module (AdaptiveBrightness.cpp/h).
- Added functions for managing ambient light sampling and brightness updates.
- Introduced a debug configuration header (debug_config.h) for easier management of debug flags.
- Updated deviceConfig to properly initialize global variables and external dependencies.
- Refactored main application logic to utilize new adaptive brightness functions and improve readability.
- Ensured consistent use of extern variables across modules for better linkage.
