+++ platformio.ini
@@ -99,6 +99,16 @@ upload_protocol = esptool
 ;  --after=hard_reset
 monitor_filters = 
 
+[env:dev-fast]
+extends = env:adafruit_matrixportal_esp32s3
+build_type = debug
+build_unflags =
+	-O2
+build_flags =
+	${env.build_flags}
+	-Og
+	-g3
+
 [env:adafruit_metro_esp32s3]
 board = adafruit_metro_esp32s3
 platform = espressif32
