+++ src/ble/ble_metrics.cpp
@@ -1,6 +1,8 @@
 #include "ble/ble_metrics.h"
 #include "ble/ble_state.h"
 
+#include "core/AdaptiveBrightness.h"
+
 #include "driver/temp_sensor.h"
 
 #include <Arduino.h>
@@ -263,7 +265,6 @@ void updateLux()
         if (deviceConnected && pLuxCharacteristic != nullptr)
         {
             // Get current lux value from the APDS9960 sensor (in lux)
-            extern uint16_t getAmbientLuxU16(); // Function from main.h
             uint16_t currentLux = getAmbientLuxU16();
 
             // Only send update if lux has changed significantly to avoid spam
